+++ packages/terra-table/src/subcomponents/Cell.jsx
@@ -15,6 +15,7 @@ import VisuallyHiddenText from 'terra-visually-hidden-text';
 
 import ColumnContext from '../utils/ColumnContext';
 import GridContext, { GridConstants } from '../utils/GridContext';
+import getFocusableElements from '../utils/focusManagement';
 import styles from './Cell.module.scss';
 
 const cx = classNames.bind(styles);
@@ -149,12 +150,7 @@ function Cell(props) {
    * Determine if cell has focusable elements
    */
   const hasFocusableElements = () => {
-    const focusableElementSelector = "a[href]:not([tabindex='-1']), area[href]:not([tabindex='-1']), input:not([disabled]):not([tabindex='-1']), "
-      + "select:not([disabled]):not([tabindex='-1']), textarea:not([disabled]):not([tabindex='-1']), button:not([disabled]):not([tabindex='-1']), "
-      + "iframe:not([tabindex='-1']), [tabindex]:not([tabindex='-1']), [contentEditable=true]:not([tabindex='-1'])";
-
-    const focusableElements = [...cellRef.current.querySelectorAll(`${focusableElementSelector}`)].filter(el => !el.hasAttribute('disabled') && !el.getAttribute('aria-hidden'));
-
+    const focusableElements = getFocusableElements(cellRef.current);
     return focusableElements.length > 0;
   };
 
@@ -177,7 +173,7 @@ function Cell(props) {
   const onMouseDown = ((event) => {
     if (!isFocusTrapEnabled) {
       onCellSelect({
-        sectionId, rowId, rowIndex, columnId, columnIndex, isShiftPressed: event.shiftKey, isCellSelectable: (!isMasked && isSelectable),
+        sectionId, rowId, rowIndex: (rowIndex - 1), columnId, columnIndex, isShiftPressed: event.shiftKey, isCellSelectable: (!isMasked && isSelectable),
       });
     }
   });
@@ -238,7 +234,7 @@ function Cell(props) {
         case KeyCode.KEY_SPACE:
           if (onCellSelect) {
             onCellSelect({
-              sectionId, rowId, rowIndex, columnId, columnIndex, isShiftPressed: event.shiftKey, isCellSelectable: (!isMasked && isSelectable),
+              sectionId, rowId, rowIndex: (rowIndex - 1), columnId, columnIndex, isShiftPressed: event.shiftKey, isCellSelectable: (!isMasked && isSelectable),
             });
           }
 
@@ -300,8 +296,8 @@ function Cell(props) {
   }
 
   // Determine table cell header attribute values
-  const sectionHeaderId = sectionId ? `${tableId}-${sectionId}` : '';
-  const rowHeaderId = !isRowHeader ? `${tableId}-rowheader-${rowId}` : '';
+  const sectionHeaderId = sectionId ? `${tableId}-${sectionId} ` : '';
+  const rowHeaderId = !isRowHeader ? `${tableId}-rowheader-${rowId} ` : '';
   const columnHeaderId = `${tableId}-${columnId}`;
 
   return (
@@ -310,7 +306,7 @@ function Cell(props) {
       ref={isGridContext ? cellRef : undefined}
       aria-selected={isSelected || undefined}
       aria-label={ariaLabel}
-      headers={`${sectionHeaderId} ${rowHeaderId} ${columnHeaderId}`}
+      headers={`${sectionHeaderId}${rowHeaderId}${columnHeaderId}`}
       tabIndex={isGridContext ? -1 : undefined}
       className={className}
       {...(isRowHeader && { scope: 'row', role: 'rowheader' })}